Move Accounting out of GCImpl.
